General error message for datasets view.

+++ hub/src/main/java/maquette/core/entities/companions/MembersCompanion.java
@@ -2,7 +2,6 @@ package maquette.core.entities.companions;
 
 import akka.Done;
 import lombok.Getter;
-import maquette.core.entities.data.datasets.exceptions.InvalidOwnerException;
 import maquette.core.ports.HasMembers;
 import maquette.core.values.ActionMetadata;
 import maquette.core.values.UID;
@@ -11,6 +10,7 @@ import maquette.core.values.authorization.GrantedAuthorization;
 import maquette.core.values.authorization.UserAuthorization;
 import maquette.core.values.data.DataAssetMemberRole;
 import maquette.core.values.exceptions.DomainException;
+import maquette.core.values.user.AuthenticatedUser;
 import maquette.core.values.user.User;
 
 import java.util.List;
@@ -35,7 +35,7 @@ public final class MembersCompanion<T extends Enum<T>> {
 
    public CompletionStage<Done> addMember(User executor, Authorization member, T role) {
       if (role.equals(DataAssetMemberRole.OWNER) && !(member instanceof UserAuthorization)) {
-         return CompletableFuture.failedFuture(InvalidOwnerException.apply());
+         return CompletableFuture.failedFuture(MembersException.invalidOwner());
       }
 
       var granted = GrantedAuthorization.apply(ActionMetadata.apply(executor), member, role);
@@ -47,10 +47,28 @@ public final class MembersCompanion<T extends Enum<T>> {
    }
 
    public CompletionStage<Done> removeMember(User executor, Authorization member) {
-      return repository.removeMember(id, member);
+      if (member instanceof UserAuthorization && executor instanceof AuthenticatedUser && member.getName().equals(((AuthenticatedUser) executor).getId())) {
+         return CompletableFuture.failedFuture(MembersException.userCannotRemoveSelf());
+      } else {
+         return repository.removeMember(id, member);
+      }
    }
 
-   public static class RemoveMemberException extends RuntimeException implements DomainException {
+   public static class MembersException extends RuntimeException implements DomainException {
+
+      private MembersException(String message) {
+         super(message);
+      }
+
+      public static MembersException userCannotRemoveSelf() {
+         var msg = "You cannot revoke your own access.";
+         return new MembersException(msg);
+      }
+
+      public static MembersException invalidOwner() {
+         var msg = "Only users are allowed to be owners of a data asset.";
+         return new MembersException(msg);
+      }
 
    }
